web: update tutorial

+++ web/install.sh
@@ -1,27 +1,27 @@
 #!/bin/bash
 
-ln -s ~/RandomMetroidSolver/web/static/favicon.ico ~/web2py/applications/solver/static/favicon.ico
-ln -s ~/RandomMetroidSolver/web/static/area_map_thumbnail.png ~/web2py/applications/solver/static/images/area_map_thumbnail.png
-ln -s ~/RandomMetroidSolver/web/static/area_map.png ~/web2py/applications/solver/static/images/area_map.png
-ln -s ~/RandomMetroidSolver/web/static/snes-controller.png ~/web2py/applications/solver/static/images/snes-controller.png
-ln -s ~/RandomMetroidSolver/web/static/ajax-loader.gif ~/web2py/applications/solver/static/images/ajax-loader.gif
-ln -s ~/RandomMetroidSolver/web/static/mystyle_20180802.css ~/web2py/applications/solver/static/css/mystyle_20180802.css
-ln -s ~/RandomMetroidSolver/web/static/bootstrap-tour.min.css ~/web2py/applications/solver/static/css/bootstrap-tour.min.css
-ln -s ~/RandomMetroidSolver/web/static/barrating ~/web2py/applications/solver/static/barrating
-ln -s ~/RandomMetroidSolver/web/static/bootstrap-tour.min.js ~/web2py/applications/solver/static/js/bootstrap-tour.min.js
-ln -s ~/RandomMetroidSolver/web/static/FileSaver.js ~/web2py/applications/solver/static/js/FileSaver.js
-ln -s ~/RandomMetroidSolver/web/static/dist ~/web2py/applications/solver/static/dist
-ln -s ~/RandomMetroidSolver/web/static/highslide ~/web2py/applications/solver/static/highslide
-ln -s ~/RandomMetroidSolver/web/views/home.html ~/web2py/applications/solver/views/solver_web/home.html
-ln -s ~/RandomMetroidSolver/web/views/presets.html ~/web2py/applications/solver/views/solver_web/presets.html
-ln -s ~/RandomMetroidSolver/web/views/solver.html ~/web2py/applications/solver/views/solver_web/solver.html
-ln -s ~/RandomMetroidSolver/web/views/randomizer.html ~/web2py/applications/solver/views/solver_web/randomizer.html
-ln -s ~/RandomMetroidSolver/web/views/infos.html ~/web2py/applications/solver/views/solver_web/infos.html
-ln -s ~/RandomMetroidSolver/web/views/stats.html ~/web2py/applications/solver/views/solver_web/stats.html
-ln -s ~/RandomMetroidSolver/web/controllers/solver_web.py ~/web2py/applications/solver/controllers/solver_web.py
-ln -s ~/RandomMetroidSolver/web/static/excellentexport.js ~/web2py/applications/solver/static/js/excellentexport.js
-ln -s ~/RandomMetroidSolver/standard_presets ~/web2py/standard_presets
-ln -s ~/RandomMetroidSolver/community_presets ~/web2py/community_presets
-ln -s ~/RandomMetroidSolver/web/static/super_metroid_logo.png ~/web2py/applications/solver/static/images/super_metroid_logo.png
-ln -s ~/RandomMetroidSolver/web/static/awesomplete.css ~/web2py/applications/solver/static/css/awesomplete.css
-ln -s ~/RandomMetroidSolver/web/static/awesomplete.js ~/web2py/applications/solver/static/js/awesomplete.js
+[ -L ~/web2py/applications/solver/static/favicon.ico ] || ln -s ~/RandomMetroidSolver/web/static/favicon.ico ~/web2py/applications/solver/static/favicon.ico
+[ -L ~/web2py/applications/solver/static/images/area_map_thumbnail.png ] || ln -s ~/RandomMetroidSolver/web/static/area_map_thumbnail.png ~/web2py/applications/solver/static/images/area_map_thumbnail.png
+[ -L ~/web2py/applications/solver/static/images/area_map.png ] || ln -s ~/RandomMetroidSolver/web/static/area_map.png ~/web2py/applications/solver/static/images/area_map.png
+[ -L ~/web2py/applications/solver/static/images/snes-controller.png ] || ln -s ~/RandomMetroidSolver/web/static/snes-controller.png ~/web2py/applications/solver/static/images/snes-controller.png
+[ -L ~/web2py/applications/solver/static/images/ajax-loader.gif ] || ln -s ~/RandomMetroidSolver/web/static/ajax-loader.gif ~/web2py/applications/solver/static/images/ajax-loader.gif
+[ -L ~/web2py/applications/solver/static/css/mystyle_20180802.css ] || ln -s ~/RandomMetroidSolver/web/static/mystyle_20180802.css ~/web2py/applications/solver/static/css/mystyle_20180802.css
+[ -L ~/web2py/applications/solver/static/css/bootstrap-tour.min.css ] || ln -s ~/RandomMetroidSolver/web/static/bootstrap-tour.min.css ~/web2py/applications/solver/static/css/bootstrap-tour.min.css
+[ -L ~/web2py/applications/solver/static/barrating ] || ln -s ~/RandomMetroidSolver/web/static/barrating ~/web2py/applications/solver/static/barrating
+[ -L ~/web2py/applications/solver/static/js/bootstrap-tour.min.js ] || ln -s ~/RandomMetroidSolver/web/static/bootstrap-tour.min.js ~/web2py/applications/solver/static/js/bootstrap-tour.min.js
+[ -L ~/web2py/applications/solver/static/js/FileSaver.js ] || ln -s ~/RandomMetroidSolver/web/static/FileSaver.js ~/web2py/applications/solver/static/js/FileSaver.js
+[ -L ~/web2py/applications/solver/static/dist ] || ln -s ~/RandomMetroidSolver/web/static/dist ~/web2py/applications/solver/static/dist
+[ -L ~/web2py/applications/solver/static/highslide ] || ln -s ~/RandomMetroidSolver/web/static/highslide ~/web2py/applications/solver/static/highslide
+[ -L ~/web2py/applications/solver/views/solver_web/home.html ] || ln -s ~/RandomMetroidSolver/web/views/home.html ~/web2py/applications/solver/views/solver_web/home.html
+[ -L ~/web2py/applications/solver/views/solver_web/presets.html ] || ln -s ~/RandomMetroidSolver/web/views/presets.html ~/web2py/applications/solver/views/solver_web/presets.html
+[ -L ~/web2py/applications/solver/views/solver_web/solver.html ] || ln -s ~/RandomMetroidSolver/web/views/solver.html ~/web2py/applications/solver/views/solver_web/solver.html
+[ -L ~/web2py/applications/solver/views/solver_web/randomizer.html ] || ln -s ~/RandomMetroidSolver/web/views/randomizer.html ~/web2py/applications/solver/views/solver_web/randomizer.html
+[ -L ~/web2py/applications/solver/views/solver_web/infos.html ] || ln -s ~/RandomMetroidSolver/web/views/infos.html ~/web2py/applications/solver/views/solver_web/infos.html
+[ -L ~/web2py/applications/solver/views/solver_web/stats.html ] || ln -s ~/RandomMetroidSolver/web/views/stats.html ~/web2py/applications/solver/views/solver_web/stats.html
+[ -L ~/web2py/applications/solver/controllers/solver_web.py ] || ln -s ~/RandomMetroidSolver/web/controllers/solver_web.py ~/web2py/applications/solver/controllers/solver_web.py
+[ -L ~/web2py/applications/solver/static/js/excellentexport.js ] || ln -s ~/RandomMetroidSolver/web/static/excellentexport.js ~/web2py/applications/solver/static/js/excellentexport.js
+[ -L ~/web2py/standard_presets ] || ln -s ~/RandomMetroidSolver/standard_presets ~/web2py/standard_presets
+[ -L ~/web2py/community_presets ] || ln -s ~/RandomMetroidSolver/community_presets ~/web2py/community_presets
+[ -L ~/web2py/applications/solver/static/images/super_metroid_logo.png ] || ln -s ~/RandomMetroidSolver/web/static/super_metroid_logo.png ~/web2py/applications/solver/static/images/super_metroid_logo.png
+[ -L ~/web2py/applications/solver/static/css/awesomplete.css ] || ln -s ~/RandomMetroidSolver/web/static/awesomplete.css ~/web2py/applications/solver/static/css/awesomplete.css
+[ -L ~/web2py/applications/solver/static/js/awesomplete.js ] || ln -s ~/RandomMetroidSolver/web/static/awesomplete.js ~/web2py/applications/solver/static/js/awesomplete.js
